Update CMakeLists.txt [decent_ci_skip]
OK, so the install(signer script could not find the launcher.sh file. It looks like it is trying to find it in the packages/Unspecified/data folder.  But I had used COMPONENT Auxiliary in the install( for the app bundle.  I am kinda hopeful that maybe this will work.  By not specifying a COMPONENT, it will end up in the Unspecified package to match the signer script.  Let's see.