Corrections to Inside Heat Balance
Two issues were noted when looking at the algorithm for inside movable insulation.  First, the information passed to CalcInteriorRadExchange should have been SurfTempInTmp not SurfTempIn since the "Tmp" version is the inside face of the movable insulation which will exchange radiation with other surfaces in the space.  Second, inside movable insulation is NOT transparent.  In develop, the impact of the SW radiation is pushed to the interface between the surface and the movable insulation.  That would be correct for transparent insulation, but E+ only allows transparent insulation at the exterior.  These corrections did NOT resolve the issue that was noted by the user though, so this is just a first step.