Solution and Unit Test
For some reason, the variable objectName was being overwritten and turned into a blank field when GetObjectItem was called.  The solution was to follow the method used by other ground temperature get routines which included moving the assignments for the base class to before the call to GetObjectItem.  The unit test shows that the name and model type are preserved and no fatal error happens.  Testing with the user example file no longer crashes but successfully completes.