Introduction of Error Traps for Missing Nodes
Boththe AirLoopHVAC:Mixer and AirLoopHVAC:Splitter do not have any sort of error message produced when the mixer has no inlet nodes or the splitter has not outlet nodes.  So, if a user by accident does not enter a mixer inlet or a splitter outlet, no error messages are produced and E+ crashes when it tries to simulate this device but the arrays are not sized to anything.  This error message would avoid a hard crash and alert the user that something is wrong (like a missing comma in the file that prompted this defect).  The new code additions here are to resolve the issue.