Fix for Missing Report Information
This corrects the issue where the exterior fenestration shading state report is not producing output for valid window shading control input.  Before this report was only being populated for windows that had a frame or divider.  Now, shading control for windows without frames/dividers will also be listed in this predefined report.