Correction of Fanger Equations
Corrected some equations where ASHRAE Standard 55/ISO 7730 have changed since the original implementation of thermal comfort in E+.  Also updated some of the unit tests to reflect the fact that the PMV and/or PPD has changed slightly as a result of the updated equations.  Might see some diffs in results, but these should be fairly minor and limited to files that report Fanger results.