Fanger PMV Fix Unit Test
This commit includes a unit test to verify that the updated version of the Fanger PMV calculation is working properly.  In addition, a statement was added to the Engineering Reference to clearly state that the method used to calculate clothing surface temperature matches the current version of ASHRAE Standard 55/ISO 7730.