Correction of Missing Design Object Issue
For inputs where there was a :Design object that contained additional data, entering nothing or a bad name for the field of the base object pointing to the :Design object led to a crash of EnergyPlus.  This was a problem for four models: Low Temp Radiant Constant Flow, Low Temp Radiant Variable Flow, Water Rad/Conv Baseboard, and Steam Rad/Conv Baseboard.  This commit fixes the issue for all four, has corrections to the IDD,  and includes some improvements/standardization to the IO Reference.