Correction for Missing Zone Control Object
Added a trap/error to avoid a hard crash when a user either enters a blank or a non-valid control name in the ZoneControl:Thermostat object.  New code avoids the hard crash when the index is zero and provides users with an error message that points toward the path for correcting the issue.