Unit Test for Fix of 11122
A unit test was added to test the presence of an invalid thermostat object.  Here, an error is purposely put in the input where a ZoneControl:Thermostat tries to access a single heating setpoint definition that does not exist.  This verifies that an error is flagged.  Before the fix, this was simply a hard crash.