Correction of Solution
After review, the changes in the first commit needed to be modified to make sure that the situation where the original problem existed was correctly singled out without changing other results that were already correction.  This commit does that and also updates some necessary changes to the new as well as other existing unit tests.