Correction of correction to time stamp
CI results came back with some differences in error messages that should not have been there.  This led to a further analysis and then correction of the code that was written to solve this issue.  This should now be all correct and it may have uncovered another subtle mistake in error reporting in the code that is now fixed.  There will potentially be some ERR file differences.