Addition of Errors for Improper Setting of Do HVAC/Plant Flags
Added code to trap for the conditions when E+ can crash.  When Do HVAC is YES but Do Plant is NO and there are no Sizing:Plant objects, a warning error is produced and Do HVAC is set to NO.  When there are no Sizing:Plant objects and both Do HVAC and Do Plant are YES, a severe/fatal error is produced.