Correction of Order Dependency Issue in AFN
The input for AirflowNetwork:Distribution:Linkage objects required that the first instance had a node that was an EnergyPlus node rather than just a node within the AFN.  Thus, the input was order dependent--something that E+ says it's not.  This fix reorders the reading of the input object so that the first one it reads is now one that has an EnergyPlus node.  The user input file that previously received a severe/fatal now runs successfully and has output that matches a version of the input file where the objects didn't violate the order issue.