Revised AFN Linkage Input Order Fix
The previous solution worked for the current version, but with the move to JSON for input, a reviewer expressed concerns that this would have to be fixed when this input was transitioned to JSON.  Since JSON reorders the input, a more comprehensive solution was needed.  This commit converts the AirflowNetwork:Distribution:Linkage input to using JSON directly and the solution should work for any input that includes these objects in any order.  So, conversion to JSON for this object is now complete as part of the solution.