Unit Test Correction
In testing the fix for this defect, it was discovered that two unit tests related to the output report tables were actually not doing anything because the unit conversion arrays were not set-up.  Fixing that, causes those two unit tests to fail as anticipated because the unit tests were still using the improper conversion factor for one particular output.  After fixing the conversion in the unit test to match the actual routine in E+, the unit tests are now working again.  So, in effect, these unit tests were actually not doing anything before and didn't fail because all of the unit conversions were coming back as 1.