Solution for Surface Level Indices not being set
The code for AFN was not setting the zone number properly for surfaces within the AFN.  Code existed for zone input to do this and link to a people statement for ASHRAE55 and CEN15251 adaptive modeling.  The correction included getting all zone and surfaces to use the same (new) subroutine, simplifying the code within the main get routine and relegating the details of linking surfaces, zones, and people statements to the new subroutine.