Correction to Defect 11171
Removed code that was checking limits because the IDD was already checking the limits properly.  Old code had an additional condtion that contradicted the IDD.  None of these limit checks are actually needed except the check making sure the lower limit is below the upper limit.  Ranges are already checked against each other and a warning produced if the ranges overlap.