Addition of Unit Test for #11171
Addition of a unit test to test the code for Defect #11171.  The problem was that the code was adding additional conditions that were unneeded and incorrect for temperatures.  Those were removed.  The unit test checks to make sure that the load range temperature max and min rules are being enforced correctly.